main/requirements.txt
langchain
faiss-cpu
sentence-transformers
streamlit
ollama
wikipedia
accelerate
sentencepiece
